requirements.txt
pandas==2.1.1
numpy==1.26.2
librosa==0.10.1
requests==2.31.0
spotipy==2.23.0
scikit-learn==1.3.0
matplotlib==3.8.2
seaborn==0.13.2
joblib==1.2.0
lightgbm==4.1.0
xgboost==1.7.3
skfeature-gli==1.1.2
yellowbrick==1.5
hdbscan==0.8.33
tqdm==4.66.1